backend/requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Background Tasks
celery==5.3.6
redis==5.0.1

# AWS S3
boto3==1.34.34

# Image Processing
Pillow==10.2.0
rembg==2.0.56
numpy==1.26.3
opencv-python-headless==4.9.0.80
scikit-image==0.22.0

# HTTP Client
httpx==0.26.0

# Utilities
python-dotenv==1.0.0
psutil==5.9.8

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

